+++ CHANGELOG.md
@@ -74,6 +74,8 @@ non-default organizations or environments are configured.
 - When installing assets, errors no longer cause file descriptors to leak, or
 lockfiles to not be cleaned up.
 - Fixed a bug where the CLI default for round robin checks was not appearing.
+- Missing custom attributes in govaluate expressions no longer result in
+an error being logged. Instead, a debug message is logged.
 
 ### Removed
 - Removed Linux/386 & Windows/386 e2e jobs on Travis CI & AppVeyor